+++ src/audio/soundcore-openal.h
@@ -33,16 +33,16 @@ typedef struct IDirectSoundBuffer ALSoundBufferStruct;
 
 typedef struct DSBUFFERDESC ALBufferParamsStruct;
 /*
-    DWORD dwSize; 
-    DWORD dwFlags; 
-    DWORD dwBufferBytes; 
-    DWORD dwReserved; 
+    unsigned long dwSize; 
+    unsigned long dwFlags; 
+    unsigned long dwBufferBytes; 
+    unsigned long dwReserved; 
     LPWAVEFORMATEX lpwfxFormat
     {
         uint16_t  wFormatTag;
         uint16_t  nChannels;
-        DWORD nSamplesPerSec;
-        DWORD nAvgBytesPerSec;
+        unsigned long nSamplesPerSec;
+        unsigned long nAvgBytesPerSec;
         uint16_t  nBlockAlign;
         uint16_t  wBitsPerSample;
         uint16_t  cbSize;
